// src/enums/NotificationType.ts
export enum NotificationType {
  CASE_CREATED = 'CASE_CREATED',
  CASE_ASSIGNED = 'CASE_ASSIGNED',
  FILE_UPLOADED = 'FILE_UPLOADED',
  CASE_COMPLETED = 'CASE_COMPLETED',
  CASE_CANCELLED = 'CASE_CANCELLED',

  PAYMENT_SUBMITTED = 'PAYMENT_SUBMITTED',
  PAYMENT_COMPLETED = 'PAYMENT_COMPLETED',
  PAYMENT_APPROVED = 'PAYMENT_APPROVED',
  PAYMENT_REJECTED = 'PAYMENT_REJECTED',


  SUBSCRIPTION_ACTIVATED = 'SUBSCRIPTION_ACTIVATED',
  SUBSCRIPTION_RENEWED = 'SUBSCRIPTION_RENEWED',
  SUBSCRIPTION_EXPIRED = 'SUBSCRIPTION_EXPIRED',
}