// src/enums/CaseStatus.ts
export enum CaseStatus {
  DRAFT = 'DRAFT',
  PENDING = 'PENDING',
  ASSIGNED = 'ASSIGNED',
  IN_PROGRESS = 'IN_PROGRESS',
  REVISION_REQUIRED = 'REVISION_REQUIRED',
  WAITING_PAYMENT = 'WAITING_PAYMENT',
  PAYMENT_APPROVED = 'PAYMENT_APPROVED',
  DELIVERED = 'DELIVERED',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED',
}